+++ shuup/default_importer/__init__.py
@@ -1,0 +1,7 @@
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+default_app_config = __name__ + ".apps.AppConfig"