+++ shuup/default_importer/importers/__init__.py
@@ -1,0 +1,11 @@
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .contact import CompanyContactImporter, PersonContactImporter
+from .product import ProductImporter
+
+__all__ = ["PersonContactImporter", "CompanyContactImporter", "ProductImporter"]