+++ doc/changelog.rst
@@ -66,6 +66,13 @@ Default Tax
 Guide
 ~~~~~
 
+Importer
+~~~~~~~~
+
+- Add Customer Importer
+- Add Product Importer
+- Add Importer
+
 Regions
 ~~~~~~~